requirements.txt
influxdb-client
python-dotenv
paho-mqtt
python-telegram-bot
psycopg2-binary
google-generativeai